docs: update frontend structure
