--- front/App.jsx
@@ -1,31 +1,0 @@
-import React, { useState } from 'react';
-import { AuthProvider, useAuth } from './context/AuthContext.jsx';
-import LoginPage from './pages/LoginPage.jsx';
-import RegisterPage from './pages/RegisterPage.jsx';
-
-function Content() {
-  const { user } = useAuth();
-  const [showRegister, setShowRegister] = useState(false);
-
-  if (!user) {
-    return showRegister ? (
-      <RegisterPage onSwitch={() => setShowRegister(false)} />
-    ) : (
-      <LoginPage onSwitch={() => setShowRegister(true)} />
-    );
-  }
-
-  return (
-    <div>
-      <h1>Welcome, {user.username}</h1>
-    </div>
-  );
-}
-
-export default function App() {
-  return (
-    <AuthProvider>
-      <Content />
-    </AuthProvider>
-  );
-}
